requirements.txt
fastapi
uvicorn
tortoise-orm
aiosqlite